Initial strong consistency prototype.
This is a very basic prototype used to benchmark the performance
impact of the on-going strong consistency research/development.

This prototype does not use the consensus library that is being
written in parallel with these changes to riak_kv. This prototype has
no leader election, is not highly available, and has several known
bugs.

All warnings aside, the client-side API (riak_client) matches the
expected behavior of the proposed work. Updates to consistent buckets
are immediate, and siblings are never returned. Likewise, updates have
conditional put behavior. A put will fail with {error, value_changed}
if the object has changed concurrently.

What is a consistent bucket? In this prototype, any bucket with a name
starting with "c#". "c#test" is a consistent bucket, "test" is a normal
Riak bucket. All consistent gets/puts use built-in quorum logic and
ignore r/w/pr/pw/etc options.

A few points:
-- This prototype uses a fixed leader per preflist. Specifically, the first
   node in the preflist.
-- Consistent read/writes are not safe during cluster changes. Don't add or
   remove nodes while performing consistent reads/writes.
-- If there is data loss on any replica, you lose all consistency guarantees.
-- There are other known, and likely unknown, bugs.