always hash using same binary format
aae via riak_object:hash/1 uses v0 format. It doesn't matter if we
hash the same bytes that are on disk since we deserialize anyways as long
as we hash the same bytes between nodes. This has the added benefit that
trees do not need to be expired if a node is downgraded and the preferred
binary version changes