Pre-loading bitcasks in parallel at app init time
This is a hacky proof of concept to work around bitcask backed vnodes
started sequentially due to supervisor serialization.  Each one tries to
load a potentially big in-memory structure from disk, so it's not
capable of saturating I/O channels. This hack adds a hook that runs at
riak_kv app initialization that pre-loads the bitcasks in parallel and
keeps those instances open until the riak_kv service is up, which
happens when all vnodes are initialized by a ring update event later.
This way, by the time the serial vnode initialization happens, opening
those bitcasks is basically a no-op, as there is an instance of the
in-memory structure to re-use.

For this to work, app config should have the following added:
    %% The hook installed
    {app_init_hooks, [{riak_kv_bitcask_pinit, run}]},
    %% Hack hack: something is initializing vnodes during the parallel
    %% load hook execution. It is not vnodes opening due to the ring
    %% update event, because that blocks until the riak_kv app is up
    %% This makes sure that vnode initialization blocks until
    %% the hook finishes and does not fail bringing down the entire
    %% node with it. TODO: figure out what is triggering this
    {open_timeout, 1000000000},
    %% How many bitcasks to load in parallel. Defaults to 4 and needs
    %% tuning to maximize disk usage.
    {bitcask_pinit, 10},